LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/form-submission-0/url-encoded-expected.txt



PASS text.simple 
PASS textarea.simple 
FAIL nokeygen.simple assert_equals: expected "foo=barb" but got "foo=barb&=MIICQTCCASkwggEiMA0GCSqGSIb3DQEBAQUAA4IBDwAwggEKAoIBAQDAKolMWpMDjDhmhvM%2BiH6BzrRtB2QiXxhpzH0PBs%2FPA%2B7DBg8Nrg33PH8BawapYzK0MZMA0biC3bnI4X%2F3y4SvgB8InNLR%2FpEE4QuTIwFrNX%2FSNZu6Bt9JSXbjKMZ073XGjUfBVotI9PfyTX4vkFoYSZllZy2wy0wDQCfNKFni5ORYHU%2FxrVnjuegGySZ80vESS65pWdgw8cUWqnWrK%2FkRb%2Bz0egAtGA274gFUuCb%2FG0EQ83%2BVEbwGNSI9NA7E4NmtVtIDXWeqhcNqCIJRKFf5wfFgVriVFxnF5rXlp9FT%2FvWZFJDf7D2GhDrdmsHQ53IyvdtNa%2FxUHIBLIKdXRWa%2FAgMBAAEWAQAwDQYJKoZIhvcNAQEEBQADggEBABkeZ0kaPY%2FzWFt6IXQ5h1eQR%2BxO8eUpe3Wr53PNydidCWK0waHqgorGDQsGV4Vaj6e%2BF2K8HZ4CDU34N710tW8T6DrJ0HWteWlaphLuoE5BecsvpB8YEkM3KbNJYPcuVlhraj177XbiPEuwFRpyqdKatxtmh%2Flj6E0eXKaKGOZKT9ICPpPSkZKDLR1jDGx5vnPNnA44IYHm75mGDQ12lXTuxLum8sUPNJVmD6iv9l4ZshWeEi6HRrCtIwp%2BfrY5r%2F4BjGt63igsZjj6QyuHHxg%2F%2FBymsO%2B4rhWUS3QVVtwT27qPobRaIvz9ZVB3WqTyHCo97W7Qa1yuIWVhALpkALw%3D"
